CB-7549 - [StatusBar][iOS 8] Landscape issue
When starting an app in landscape mode in ios 8 and then calling StatusBar.overlaysWebView(false), the screen would be pushed down all the way.  Changing the orientation did not fix the issue, leaving the page pushed down the width of the device.

We found that excluding the changes above in ios 8 fixed the issue, although it's possible that this is hiding a landscape orientation bug in ios 8 that may get fixed later.

Signed-off-by: Shazron Abdullah <shazron@apache.org>
(cherry picked from commit 921daccd2cc7be5294488adb64b8c4c36363b63c)